apps/users/login_backend.py
from django.contrib.auth import get_user_model

UserModel = get_user_model()


class HoloModelBackend:
    def authenticate(self, request, username=None, password=None, **kwargs):
        if '@' in username:
            kwargs = {'email__iexact': username}
        else:
            kwargs = {'phone': username}
        try:
            user = UserModel.objects.get(**kwargs)
            if user.check_password(password):
                return user
        except UserModel.DoesNotExist:
            return None

    def get_user(self, user_id):
        try:
            return UserModel.objects.get(pk=user_id)
        except UserModel.DoesNotExist:
            return None
